refactor: searchbar input component
